+++ chunky/src/java/se/llbit/chunky/renderer/postprocessing/HableToneMappingFilter.java
@@ -1,4 +1,4 @@
-package se.llbit.chunky.renderer.postprecessing;
+package se.llbit.chunky.renderer.postprocessing;
 
 /**
  * Implementation of Hable tone mapping
