+++ chunky/src/java/se/llbit/chunky/renderer/postprocessing/PreviewFilter.java
@@ -1,4 +1,4 @@
-package se.llbit.chunky.renderer.postprecessing;
+package se.llbit.chunky.renderer.postprocessing;
 
 import org.apache.commons.math3.util.FastMath;
