+++ chunky/src/java/se/llbit/chunky/renderer/postprocessing/Tonemap1Filter.java
@@ -1,4 +1,4 @@
-package se.llbit.chunky.renderer.postprecessing;
+package se.llbit.chunky.renderer.postprocessing;
 
 import se.llbit.math.QuickMath;
