+++ chunky/src/java/se/llbit/chunky/renderer/scene/Scene.java
@@ -71,10 +71,9 @@ import se.llbit.chunky.renderer.ResetReason;
 import se.llbit.chunky.renderer.WorkerState;
 import se.llbit.chunky.renderer.export.PictureExportFormat;
 import se.llbit.chunky.renderer.projection.ProjectionMode;
-import se.llbit.chunky.renderer.postprecessing.GammaCorrectionFilter;
-import se.llbit.chunky.renderer.postprecessing.PostProcessingFilter;
-import se.llbit.chunky.renderer.postprecessing.PostProcessingFilters;
-import se.llbit.chunky.renderer.postprecessing.PreviewFilter;
+import se.llbit.chunky.renderer.postprocessing.PostProcessingFilter;
+import se.llbit.chunky.renderer.postprocessing.PostProcessingFilters;
+import se.llbit.chunky.renderer.postprocessing.PreviewFilter;
 import se.llbit.chunky.renderer.renderdump.RenderDump;
 import se.llbit.chunky.resources.BitmapImage;
 import se.llbit.chunky.resources.OctreeFileFormat;
