+++ chunky/src/java/se/llbit/chunky/ui/render/PostprocessingTab.java
@@ -23,12 +23,8 @@ import javafx.scene.Node;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.ScrollPane;
 import javafx.util.StringConverter;
-import se.llbit.chunky.renderer.Postprocess;
-import se.llbit.chunky.renderer.RenderMode;
-import se.llbit.chunky.renderer.export.PictureExportFormat;
-import se.llbit.chunky.renderer.export.PictureExportFormats;
-import se.llbit.chunky.renderer.postprecessing.PostProcessingFilter;
-import se.llbit.chunky.renderer.postprecessing.PostProcessingFilters;
+import se.llbit.chunky.renderer.postprocessing.PostProcessingFilter;
+import se.llbit.chunky.renderer.postprocessing.PostProcessingFilters;
 import se.llbit.chunky.renderer.scene.Scene;
 import se.llbit.chunky.ui.DoubleAdjuster;
 import se.llbit.chunky.ui.RenderControlsFxController;
